fix: PCI 64-bit BAR support — combines BAR[n] and BAR[n+1] for addresses above 4GB
pci_get_bar_address() now returns uint64_t. Detects 64-bit memory BARs
(type bits = 0x04) and combines the low BAR with BAR[n+1] for the upper
32 bits. Required for PCIe devices with BARs above 4GB (RTX 3060, NVMe).

Added pci_is_bar_64bit() helper. Updated all callers to uint64_t.
4 new mock tests with simulated GPU device at 0x1_FC000000.

15/15 PCI tests PASS, 6/6 NIC tests PASS.

+++ phase3/src/drivers/nic_rtl8168.c
@@ -285,7 +285,7 @@ int rtl_nic_probe_pci(rtl_nic_t *nic)
 {
     pci_device_t devices[32];
     int count, i;
-    uint32_t bar_addr;
+    uint64_t bar_addr;
 
     if (!nic)
         return -1;
